can now delete vpc

+++ aws.py
@@ -1,7 +1,7 @@
 from ConfigParser import SafeConfigParser
 from boto.s3.connection import Location
 from collections import namedtuple
-import os, boto, boto.ec2
+import os, boto, boto.ec2, time
 
 CREDENTIALS_FILE = os.path.join(os.path.dirname(__file__), 'venv', 'aws.cfg')
 BASTION_KEY_FILE = os.path.join(os.path.dirname(__file__), 'venv', 'bastion.pem')
@@ -9,17 +9,26 @@ BASTION_KEY_FILE = os.path.join(os.path.dirname(__file__), 'venv', 'bastion.pem'
 ANYWHERE = '0.0.0.0/0'
 
 Node = namedtuple('Node', ['name', 'public_ip', 'user', 'keyfile'])
+Connections = namedtuple('Connections', ['ec2', 'vpc', 's3'])
+
+def connect(vpc_region):
+    access_key_id, secret_access_key = read_credentials()
+    s3_conn = boto.connect_s3(access_key_id, secret_access_key)
+    ec2_conn = boto.ec2.connect_to_region(vpc_region, aws_access_key_id=access_key_id, aws_secret_access_key=secret_access_key)
+    region = boto.ec2.get_region(vpc_region, aws_access_key_id=access_key_id, aws_secret_access_key=secret_access_key)
+    vpc_conn = boto.connect_vpc(access_key_id, secret_access_key, region=region)
+    return Connections(ec2_conn, vpc_conn, s3_conn)
 
 def make_vpc(vpc_config_name):
     vpc_config = read_vpc_config(vpc_config_name)
 
     vpc_name = vpc_config.get('vpc', 'name')
     vpc_region = vpc_config.get('vpc', 'region')
+    connections = connect(vpc_region)
 
-    access_key_id, secret_access_key = read_credentials()
-    vpc_conn = boto.connect_vpc(access_key_id, secret_access_key, region=vpc_region)
-    ec2_conn = boto.ec2.connect_to_region(vpc_region, aws_access_key_id=access_key, aws_secret_access_key=secret_access_key)
-    s3_conn = boto.connect_s3(access_key_id, secret_access_key)
+    ec2_conn = connections.ec2
+    vpc_conn = connections.vpc
+    s3_conn = connections.s3
 
     vpc = get_or_create_vpc(vpc_conn, vpc_name, vpc_config.get('vpc', 'cidr_block'))
     internet_gateway = get_or_create_internet_gateway(vpc_conn, vpc, vpc_name)
@@ -64,7 +73,7 @@ def get_or_create_route_table(vpc_conn, vpc, route_name, internet_gateway):
     print 'Creating Route Table with name:', route_name
     route_table = vpc_conn.create_route_table(vpc.id)
     tag_with_name(route_table, route_name)
-    vpc_conn.create_route(route_table.id, ANYWHERE, internet_gateway)
+    vpc_conn.create_route(route_table.id, ANYWHERE, internet_gateway.id)
     return route_table
 
 def get_or_create_subnet(vpc_conn, vpc, route_table, subnet_name, cidr_block, availability_zone):
@@ -81,34 +90,40 @@ def get_or_create_bastion_host(ec2_conn, s3_conn, vpc_config, bastion_host_name,
     instance_type = vpc_config.get('vpc', 'default_instance_type')
     image_login_user = vpc_config.get('vpc', 'default_image_login_user')
     key_pair = get_bastion_host_key(ec2_conn, s3_conn, vpc_config)
-    security_group = get_or_create_vpc_security_group(ec2_conn, vpc_id)
+    security_group = get_or_create_vpc_security_group(ec2_conn, vpc_config, vpc_id)
 
     for reservation in fetch_running_reservations(ec2_conn, bastion_host_name):
         for instance in reservation.instances:
-            return Node(bastion_host_name, instance.ip_address, image_login_user, BASTION_KEY_FILE)
+            public_ip = associate_elastic_ip(ec2_conn, instance)
+            return Node(bastion_host_name, public_ip, image_login_user, BASTION_KEY_FILE)
 
     print 'Creating Instance with name:', bastion_host_name
     reservation = ec2_conn.run_instances(image_id, instance_type=instance_type,
-                        key_name=key_pair.name, security_group_ids=[security_group.id],
-                        subnet_id=subnet.id)
+                    key_name=key_pair.name, security_group_ids=[security_group.id],
+                    subnet_id=subnet.id)
 
     for instance in reservation.instances:
         print 'Waiting for', bastion_host_name, instance.id, 'to start ...'
         wait_until(instance, 'running')
         tag_with_name(instance, bastion_host_name)
-        elastic_ip = get_or_create_elastic_ip(ec2_conn)
-        ec2_conn.associate_address(instance.id, elastic_ip)
-        print bastion_host_name, 'is associated with Elastic IP', elastic_ip
-        return Node(bastion_host_name, elastic_ip, image_login_user, BASTION_KEY_FILE)
+        public_ip = associate_elastic_ip(ec2_conn, instance)
+        print bastion_host_name, 'is associated with Elastic IP', public_ip
+        return Node(bastion_host_name, public_ip, image_login_user, BASTION_KEY_FILE)
+
+def associate_elastic_ip(ec2_conn, instance):
+    if instance.ip_address:
+        return instance.ip_address
+    elastic_ip = get_or_create_elastic_ip(ec2_conn)
+    ec2_conn.associate_address(instance_id=instance.id, allocation_id=elastic_ip.allocation_id)
+    return elastic_ip.public_ip
 
 def get_or_create_elastic_ip(ec2_conn):
     filters = {'domain': 'vpc'}
     for address in ec2_conn.get_all_addresses(filters=filters):
         if not address.instance_id:
-            return address.public_ip
+            return address
     print 'Creating a new Elastic IP'
-    address = ec2_conn.allocate_address('vpc')
-    return address.public_ip
+    return ec2_conn.allocate_address('vpc')
 
 def get_bastion_host_key(ec2_conn, s3_conn, vpc_config):
     vpc_name = vpc_config.get('vpc', 'name')
@@ -133,15 +148,15 @@ def write_bastion_key_file(contents):
     os.chmod(BASTION_KEY_FILE, 0600)
 
 def upload_bastion_key(s3_conn, vpc_config, key_object_name):
-    print 'Uploading bastion host key to S3 bucket'
     bucket = get_key_bucket(s3_conn, vpc_config)
+    print 'Uploading bastion host key to S3 bucket'
     key = bucket.new_key(key_object_name)
     key.set_contents_from_filename(BASTION_KEY_FILE, encrypt_key=True)
 
 def ensure_bastion_host_keyfile_exists(s3_conn, vpc_config, key_object_name):
     if not os.path.isfile(BASTION_KEY_FILE):
-        print 'Downloading bastion host key from S3 bucket'
         bucket = get_key_bucket(s3_conn, vpc_config)
+        print 'Downloading bastion host key from S3 bucket'
         key = bucket.get_key(key_object_name)
         key.get_contents_to_filename(BASTION_KEY_FILE)
         os.chmod(BASTION_KEY_FILE, 0600)
@@ -149,7 +164,7 @@ def ensure_bastion_host_keyfile_exists(s3_conn, vpc_config, key_object_name):
 def get_key_bucket(s3_conn, vpc_config):
     key_bucket_region = get_or_default(vpc_config, 'vpc', 'key_bucket_region', Location.DEFAULT)
     key_bucket_prefix = vpc_config.get('vpc', 'key_bucket_prefix')
-    bucket_name = (key_bucket_prefix + s3_conn.aws_access_key_id).lower()
+    bucket_name = (key_bucket_prefix + '-' + s3_conn.aws_access_key_id).lower()
     print 'Using', bucket_name, 'to hold EC2 keys'
     return s3_conn.create_bucket(bucket_name, location=key_bucket_region)
 
@@ -160,11 +175,12 @@ def get_or_create_vpc_security_group(ec2_conn, vpc_config, vpc_id):
     vpc_name = vpc_config.get('vpc', 'name')
     security_group_name = vpc_name + '-bastion'
 
-    for security_group in ec2_conn.get_all_security_groups([security_group_name]):
+    filters = {'group-name': security_group_name}
+    for security_group in ec2_conn.get_all_security_groups(filters=filters):
         return security_group
 
     print 'Creating Security Group with name:', security_group_name
-    security_group ec2_conn.create_security_group(security_group_name, security_group_name, vpc_id)
+    security_group = ec2_conn.create_security_group(security_group_name, security_group_name, vpc_id)
     clear_all_permissions(ec2_conn, security_group) # start with a clean slate
     allow_https_egress(ec2_conn, security_group.id, ANYWHERE)
     allow_http_egress(ec2_conn, security_group.id, ANYWHERE)
@@ -177,7 +193,7 @@ def allow_ssh_ingress(ec2_conn, security_group, source):
 def allow_http_egress(ec2_conn, security_group, destination):
     ec2_conn.authorize_security_group_egress(security_group, 'tcp', 80, 80, None, destination)
 
-def allow_https_egress(security_group, destination):
+def allow_https_egress(ec2_conn, security_group, destination):
     ec2_conn.authorize_security_group_egress(security_group, 'tcp', 443, 443, None, destination)
 
 def fetch_running_reservations(ec2_conn, name):
@@ -221,7 +237,7 @@ def read_credentials():
 
 def read_config_file(config_file_path):
     config = SafeConfigParser()
-    with open(vpc_config_file) as fp:
+    with open(config_file_path) as fp:
         config.readfp(fp)
     return config
 
@@ -230,14 +246,17 @@ def delete_vpc(vpc_config_name):
 
     vpc_name = vpc_config.get('vpc', 'name')
     vpc_region = vpc_config.get('vpc', 'region')
+    connections = connect(vpc_region)
 
-    access_key_id, secret_access_key = read_credentials()
-    vpc_conn = boto.connect_vpc(access_key_id, secret_access_key, region=vpc_region)
-    ec2_conn = boto.ec2.connect_to_region(vpc_region, aws_access_key_id=access_key, aws_secret_access_key=secret_access_key)
+    ec2_conn = connections.ec2
+    vpc_conn = connections.vpc
 
     for vpc in filter_by_name(vpc_conn.get_all_vpcs, vpc_name):
+        instances = []
         filters = {'vpc-id': vpc.id}
-        instances = [instance for instance in ec2_conn.get_all_instances(filters=filters)]
+        for reservation in ec2_conn.get_all_instances(filters=filters):
+            for instance in reservation.instances:
+                instances.append(instance)
 
         filters = {'instance-id': [instance.id for instance in instances]}
         addresses = [address for address in ec2_conn.get_all_addresses(filters=filters)]
@@ -259,14 +278,18 @@ def delete_vpc(vpc_config_name):
             wait_until(instance, 'terminated')
 
         # there is no filter to get security groups for a vpc!!
-        security_groups = [group for group in ec2_conn.get_all_security_groups() if group.vpc_id == vpc.id]
+        security_groups = [gp for gp in ec2_conn.get_all_security_groups() if gp.vpc_id == vpc.id]
 
         for security_group in security_groups:
             clear_all_permissions(vpc_conn, security_group)
 
+        if len(security_groups) > 1:
+            time.sleep(5)
+
         for security_group in security_groups:
-            print 'Deleting security group', security_group.name
-            security_group.delete()
+            if security_group.name != 'default':
+                print 'Deleting security group', security_group.name
+                security_group.delete()
 
         filters = {'vpc-id': vpc.id}
 
@@ -274,20 +297,22 @@ def delete_vpc(vpc_config_name):
             print 'Deleting subnet', subnet.id
             vpc_conn.delete_subnet(subnet.id)
 
-        for route_table in vpc.get_all_route_tables(filters=filters):
-            print 'Deleting route table', route_table.id
-            vpc_conn.delete_route_table(route_table.id)
+        for route_table in vpc_conn.get_all_route_tables(filters=filters):
+            if not is_main_route_table(route_table):
+                print 'Deleting route table', route_table.id
+                vpc_conn.delete_route_table(route_table.id)
 
         for internet_gateway in filter_by_name(vpc_conn.get_all_internet_gateways, vpc_name):
             print 'Deleting internet gateway', internet_gateway.id
+            vpc_conn.detach_internet_gateway(internet_gateway.id, vpc.id)
             vpc_conn.delete_internet_gateway(internet_gateway.id)
 
         print 'Deleting vpc', vpc_name, vpc.id
         vpc_conn.delete_vpc(vpc.id)
 
 def clear_all_permissions(ec2_conn, security_group):
-    revoke_ingress_permissions(vpc_conn, security_group)
-    revoke_egress_permissions(vpc_conn, security_group)
+    revoke_ingress_permissions(ec2_conn, security_group)
+    revoke_egress_permissions(ec2_conn, security_group)
 
 def revoke_ingress_permissions(ec2_conn, security_group):
     for rule in security_group.rules:
@@ -301,3 +326,8 @@ def revoke_egress_permissions(ec2_conn, security_group):
         for grant in rule.grants:
             ec2_conn.revoke_security_group_egress(security_group.id, rule.ip_protocol,
                 rule.from_port, rule.to_port, grant.group_id, grant.cidr_ip)
+
+def is_main_route_table(route_table):
+    for association in route_table.associations:
+        if association.main:
+            return True
